+++ client/templates/logic/movieInfo.js
@@ -279,7 +279,7 @@ function setArrayMovieInfoBoxes(data, dataType) {
         _.each(data.statuses, function(ele) {
             arrayMovieInfoBoxes.push({
                 boxType: 'boxTweet',
-                background: 'background-color: #FAFAFA',
+                background: 'background-color: #FFFFFF', // #FAFAFA
                 isUser: true,
                 isTwitter: true,
                 user: ele.user.screen_name,
@@ -324,7 +324,7 @@ function setArrayMovieInfoBoxes(data, dataType) {
             i++;
             arrayMovieInfoBoxes.push({
                 boxType: 'boxReview',
-                background: 'background-color: #FAFAFA',
+                background: 'background-color: #FFFFFF', // #FAFAFA
                 isUser: true,
                 isTwitter: false,
                 user: ele.author + ' - TMDb',
@@ -342,18 +342,18 @@ function setArrayMovieInfoBoxes(data, dataType) {
             shuffle(arrayMovieInfoBoxes, 1);
             return;
         }
-        _.each(ris.result, function(ele) {
+        _.each(ris.results, function(ele) {
             if (i === 10)
                 return;
             i++;
             stars += parseInt(ele.score, 10);
             arrayMovieInfoBoxes.push({
                 boxType: 'boxReview',
-                background: 'background-color: #FAFAFA',
+                background: 'background-color: #FFFFFF', // #FAFAFA
                 isUser: true,
                 isTwitter: false,
                 user: ele.critic,
-                text: ele.excerpt,
+                text: '“' + ele.excerpt,
                 link: ele.link
             });