+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Film Hunter
 *Changelog*
 
+**[0.5.5]** *24/04/2015*
+- Changed some elements of the UI;
+- Implemented public accounts;
+- Implemented custom account url;
+- SEO optimizations;
+- Updated Mashape API;
+- Implemented code against XSS;
+- Improved code;
+- Various bugfix.
+
 **[0.5.4]** *17/04/2015*
 - New account page;
 - Added login with Twitter and Facebook;